Finished and Tested API

+++ src/main/java/com/prokarma/service/CustomerService.java
@@ -1,13 +1,10 @@
 package com.prokarma.service;
 
-import java.util.ArrayList;
-import java.util.List;
 
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.prokarma.model.Customer;
-import com.prokarma.model.Vehicle;
 import com.prokarma.repository.CustomerRepository;
 import com.prokarma.repository.VehicleRepository;
 
@@ -25,51 +22,24 @@ public class CustomerService {
 	}
 
 	public List<Customer> getAll() {
-		List<Customer> customers = customerRepository.findAll();
-		List<Vehicle> vehicles = new ArrayList<>();
-		vehicles = vehicleRepository.findAll();
-		for (Customer customer : customers) {
-			for (Vehicle vehicle : vehicles) {
-				if (vehicle.getOwnerId() == customer.getId()) {
-					customer.addVehicle(vehicle);
-				}
-			}
-		}
-		
-		return customers;
+		return customerRepository.findAll();
 	}
-/*
+
 	public Customer getCustomerById(Integer id) {
-		Customer customer = customerRepository.getOne(id);
-		List<Vehicle> vehicles = vehicleRepository.findAll();
-		for (Vehicle vehicle : vehicles) {
-			if (vehicle.getOwner_id() == customer.getId()) {
-				customer.addVehicle(vehicle);
-			}
-		}
-		return customer;
+		return customerRepository.getById(id);
 	}
 	
 	public Customer getCustomerByName(String name) {
-		Customer customer = customerRepository.getCustomerByName(name);
-		List<Vehicle> vehicles = vehicleRepository.findAll();
-		for (Vehicle vehicle : vehicles) {
-			if (vehicle.getOwner_id() == customer.getId()) {
-				customer.addVehicle(vehicle);
-			}
-		}
-		return customer;
+		return customerRepository.getCustomerByName(name);
 	}
 	
-	public void deleteCustomer(Integer id) {
-		customerRepository.deleteById(id);
-		List<Vehicle> vehicles = vehicleRepository.findAll();
-		for (Vehicle vehicle : vehicles) {
-			if (vehicle.getOwner_id() == id) {
-				vehicles.remove(vehicle);
-			}
+	public Boolean deleteCustomer(Integer id) {
+		Customer customer = customerRepository.getOne(id);
+		if (customer != null && customer.getVehiclesOwned().isEmpty()) {
+			customerRepository.deleteById(id);
+			return true;
 		}
-		vehicleRepository.saveAll(vehicles);
+		return false;
 	}
-	*/
+	
 }
